ok i pull up

+++ projet_zeldiablo/src/gameLaby/laby/LabyDessin.java
@@ -26,7 +26,7 @@ public class LabyDessin implements DessinJeu {
 
                 if (laby.pj.etrePresent(i, j)) {
                     gc.setFill(Color.RED);
-                    gc.setGlobalAlpha(laby.pj.getVie()/laby.pj.getMaxVie());
+                    gc.setGlobalAlpha(((double)(laby.pj.getVie())/(laby.pj.getMaxVie())));
                     gc.fillOval(i * taille, j * taille, taille, taille);
                 }
